requirements.txt
click==7.1.2
Flask==1.1.2
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
mysql-connector-python==8.0.27
PyMySQL==1.0.2
Werkzeug==1.0.1
WTForms==3.0.0
Flask-WTF==0.15.1